+++ src/components/SSD/ParadigmBlocks/Assembled.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
     import { Canvg } from "canvg";
-    import { afterUpdate, beforeUpdate, onMount } from 'svelte';
+    import { afterUpdate, beforeUpdate, onMount, createEventDispatcher } from 'svelte';
     import { ship } from "@/stores/writeShip";
     import type { ILayout as ILayoutSystem } from "@/stores/writeShip";
     import { ssdComponents } from "@/stores/writeSsd";
@@ -9,6 +9,8 @@
     import { genSvg } from "@/lib/hull";
     import type { ISystem } from "src/lib/systems/_base";
     import type { ISystemSVG } from "@/lib/svgLib";
+
+    const dispatch = createEventDispatcher();
 
     let layout = ($ship.layout as ILayoutSystem).blocks;
     const svgCore = svgLib.find(x => x.id === "coreSys")!;
@@ -35,18 +37,7 @@
     let svgDataStr: string;
     let pngDataStr: string;
     let pngCanvas: HTMLCanvasElement;
-    $: if (fullSsdSvg !== undefined) {
-        svgDataStr = "data:image/svg+xml;charset=utf-8," + encodeURIComponent(fullSsdSvg.outerHTML.replaceAll(`href=`, `xlink:href=`));
-
-        genPng();
-    }
-
-    const genPng = () => {
-        const ctx = pngCanvas.getContext("2d");
-        const v = Canvg.fromString(ctx, fullSsdSvg.outerHTML);
-        v.render();
-        pngDataStr = pngCanvas.toDataURL("image/png");
-    };
+    let injectXlink: boolean;
 
     const genHull = () => {
         $ssdComponents.hull = genSvg($ship, layout.blocks.cellsize, {height: layout.blocks.blockHull.height, width: layout.blocks.blockHull.width});
@@ -86,6 +77,24 @@
     let coreHeightOffset = 0;
     let coreOffsetFactor = 0.15;
     afterUpdate(() => {
+        if (fullSsdSvg !== undefined) {
+            let text: string;
+            if (injectXlink) {
+                text = fullSsdSvg.outerHTML;
+                text = text.replaceAll(`href=`, `xlink:href=`);
+
+            } else {
+                text = fullSsdSvg.outerHTML;
+                text = text.replace(`<svg `, `<?xml version="1.0" standalone="no"?><!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd"><svg version="1.1" xmlns="http://www.w3.org/2000/svg" `);
+            }
+            svgDataStr = "data:image/svg+xml;charset=utf-8," + encodeURIComponent(text);
+
+            const ctx = pngCanvas.getContext("2d");
+            const v = Canvg.fromString(ctx, fullSsdSvg.outerHTML);
+            v.render();
+            pngDataStr = pngCanvas.toDataURL("image/png");
+        }
+
         coreHeightOffset = (layout.blocks.blockCore.height * coreOffsetFactor) / 2;
         coreWidthOffset = (layout.blocks.blockCore.width * coreOffsetFactor) / 2;
     });
@@ -124,9 +133,20 @@
 
 <div class="level">
     <div class="level-item">
-        <a href="{svgDataStr}" download="SSD.svg">
-            <button class="button">Download SVG</button>
-        </a>
+        <div class="field">
+            <div class="control">
+            <a href="{svgDataStr}" download="SSD.svg">
+                <button class="button">Download SVG</button>
+            </a>
+            </div>
+            <div class="control">
+                <label class="checkbox">
+                    <input type="checkbox" bind:checked="{injectXlink}">
+                    Adjust for apps (<a on:click="{() => dispatch("message", {msg: "showSvg"})
+                    }">read more</a>)
+                  </label>
+            </div>
+        </div>
     </div>
     <div class="level-item">
         <a href="{pngDataStr}" download="SSD.png">